requirements.txt
flask
requests
google-generativeai
python-dotenv
pytube
youtube-transcript-api